[refactor] input out of the model.
